% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_plot.R
\name{diff_plot}
\alias{diff_plot}
\title{Plot of Region and State HPI Difference}
\usage{
diff_plot(State)
}
\arguments{
\item{State}{Character.
A state name from hpi dataset,
you can choose from unique(yang7er::hpi$state)}
}
\value{
A plot.
}
\description{
Show a plot of the difference of hpi between region and state
}
\details{
This function using data 'hpi', ploting the difference of hpi
between region and state. You can put a state name.
It will draw a comparative graph of the HPI changes in your chosen state,
and the region where the state is located
}
\examples{
diff_plot("NY")


}
