% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_region_state.R
\name{diff_region_state}
\alias{diff_region_state}
\title{Region and State HPI Difference}
\usage{
diff_region_state(State, Year)
}
\arguments{
\item{State}{Character.
A state name from hpi dataset,
you can choose from unique(yang7er::hpi$state)}

\item{Year}{Numeric.
A year from 1975 to 2018,
you can choose from unique(yang7er::hpi$year)}
}
\value{
A data.frame. The information include:
\itemize{
\item \code{year} : The year you chosen.
\item \code{state} : The state you chosen.
\item \code{region} : The region to which your chosen state belongs
\item \code{region avg} : Region average HPI.
\item \code{state avg} : State average HPI.
\item \code{diff} : Region average HPI- State average HPI.
}
}
\description{
Calculating the difference of hpi between region and state
}
\details{
This function using data 'hpi', calculating the difference of hpi
between region and state. You can put a state name, and  a year,
it will show a tibble include year, state, region, region avg, state avg and
diff.
}
\examples{
diff_region_state("AK",2018)

}
