% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hpi.R
\name{get_hpi}
\alias{get_hpi}
\title{Get data HPI}
\format{
A data frame with 26877 rows and 9 variables:
\itemize{
\item \code{year}:year
\item \code{month}:month
\item \code{state}:states in US
\item \code{price_index}:house price index
\item \code{us_avg}:us annual average HPI
\item \code{year_avg}:every stats's anaual average HPI
\item \code{us_year_avg}:us annual average HPI
\item \code{region}:state belongs in which region
\item \code{region_avg}:every region's anaual average HPI
}
}
\source{
\url{https://raw.githubusercontent.com/rfordatascience/tidytuesday/master/data/2019/2019-02-05/state_hpi.csv"}
}
\usage{
get_hpi()
}
\description{
the function of get the HPI data
}
\details{
A dataset containing the year,month,state,price_index,us_avg,year_avg,us_year_avg,region,region_avg
}
\examples{
get_hpi
}
