% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hpi}
\alias{hpi}
\title{HPI data in US from 1975-2018}
\format{
A data frame with 26877 rows and 9 variables:
\describe{
\item{year}{year}
\item{month}{month}
\item{state}{states in US}
\item{price_index}{house price index}
\item{us_avg}{us annual average HPI}
\item{year_avg}{every stats's anaual average HPI}
\item{us_year_avg}{us annual average HPI}
\item{region}{state belongs in which region}
\item{region_avg}{every region's anaual average HPI}
}
}
\source{
\url{https://www.fhfa.gov/DataTools/Downloads/Pages/House-Price-Index.aspx}
}
\usage{
hpi
}
\description{
A dataset containing the year,month,state,price_index,us_avg,year_avg,us_year_avg,region,region_avg
}
\keyword{datasets}
