% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions_plot.R
\name{regions_plot}
\alias{regions_plot}
\title{Plot of Regions HPI Difference}
\usage{
regions_plot(Min, Max)
}
\arguments{
\item{Min, Number.}{A min year (the plot beginning from which year) you chose
you can choose from unique(yang7er::hpi$year)}

\item{Max, Number.}{A max year (the plot ending from which year) you chose
you can choose from unique(yang7er::hpi$year)}
}
\value{
A plot.
}
\description{
Show a plot of the hpi change between in different regions
}
\details{
This function using data 'hpi', ploting the hpi changes
in all regions. You can put a year. like(1990,2018)
It will draw a graph of the HPI changes in your chosen years for
all regions.
}
\examples{
#Plot from 1995 to 2017
regions_plot(1995,2017)

}
